# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Leave & Working Days API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  user: 'user',
  admin: 'admin',
} as const;

export type LeaveType = typeof LeaveType[keyof typeof LeaveType];


export const LeaveType = {
  annual: 'annual',
  sick: 'sick',
} as const;

export type RequestStatus = typeof RequestStatus[keyof typeof RequestStatus];


export const RequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  cancelled: 'cancelled',
} as const;

export type DayStatus = typeof DayStatus[keyof typeof DayStatus];


export const DayStatus = {
  working: 'working',
  non_working: 'non_working',
  annual_pending: 'annual_pending',
  annual_approved: 'annual_approved',
  annual_rejected: 'annual_rejected',
  sick_pending: 'sick_pending',
  sick_approved: 'sick_approved',
  sick_rejected: 'sick_rejected',
} as const;

export interface User {
  id: number;
  name: string;
  firstName: string;
  lastName: string;
  email: string;
  role: Role;
  startDate: string;
  active: boolean;
  annualEntitlement: number;
  sickEntitlement: number;
  balanceAdjustment?: number;
  createdAt: string;
  updatedAt: string;
}

export interface Balance {
  leaveYear: number;
  annualEntitlement: number;
  accrued: number;
  carriedOver: number;
  /** @nullable */
  carryOverDeadline: string | null;
  usedPaid: number;
  usedUnpaid: number;
  available: number;
  pending: number;
  wouldBeUnpaid: number;
  sickUsed: number;
  sickRemaining: number;
  balanceAdjustment: number;
}

export interface UserWithBalance {
  user: User;
  balance: Balance;
}

export interface ApprovedLeavePeriod {
  id: number;
  type: LeaveType;
  startDate: string;
  endDate: string;
  workingDays: number;
}

export interface Dashboard {
  user: User;
  balance: Balance;
  nextApprovedLeave: ApprovedLeavePeriod[];
  sickOverAllowance: boolean;
}

export interface CalendarDay {
  id: number;
  userId: number;
  date: string;
  status: DayStatus;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  requestId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CalendarMarkInput {
  dates: string[];
  status: DayStatus;
  note?: string;
}

export interface LeaveRequest {
  id: number;
  userId: number;
  type: LeaveType;
  startDate: string;
  endDate: string;
  workingDays: number;
  paidDays: number;
  unpaidDays: number;
  status: RequestStatus;
  /** @nullable */
  userNote?: string | null;
  /** @nullable */
  adminNote?: string | null;
  /** @nullable */
  approvedBy?: number | null;
  /** @nullable */
  approvedAt?: string | null;
  /** @nullable */
  rejectedBy?: number | null;
  /** @nullable */
  rejectedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AdminLeaveRequest {
  request: LeaveRequest;
  userName: string;
  userEmail: string;
  accruedAtDate: number;
  usedAnnual: number;
  sickUsed?: number;
  sickOverAllowance?: boolean;
}

export interface LeaveRequestInput {
  type: LeaveType;
  startDate: string;
  endDate: string;
  userNote?: string;
}

export interface LeavePreview {
  workingDays: number;
  paidDays: number;
  unpaidDays: number;
  hasUnpaid: boolean;
  availableBeforeRequest: number;
  sickOverAllowance?: boolean;
}

export interface ReviewInput {
  adminNote?: string;
}

export interface ProfileUpdate {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
}

export interface UserInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  email: string;
  role?: Role;
  startDate: string;
  annualEntitlement?: number;
  sickEntitlement?: number;
}

export interface UserUpdate {
  /** @minLength 1 */
  firstName?: string;
  /** @minLength 1 */
  lastName?: string;
  email?: string;
  role?: Role;
  startDate?: string;
  active?: boolean;
  annualEntitlement?: number;
  sickEntitlement?: number;
}

export interface BalanceOverrideInput {
  adjustment: number;
  note?: string;
}

export interface Settings {
  id: number;
  annualEntitlement: number;
  sickEntitlement: number;
  leaveYearStartMonth: number;
  leaveYearStartDay: number;
  accrualDay: number;
  maxRolloverDays: number;
  carryOverDeadlineMonths: number;
  allowExceedBalance: boolean;
  excessBecomesUnpaid: boolean;
  autoWeekends: boolean;
  allowHalfDays: boolean;
  emailNotifications: boolean;
  updatedAt: string;
}

export interface SettingsUpdate {
  annualEntitlement?: number;
  sickEntitlement?: number;
  leaveYearStartMonth?: number;
  leaveYearStartDay?: number;
  accrualDay?: number;
  maxRolloverDays?: number;
  carryOverDeadlineMonths?: number;
  allowExceedBalance?: boolean;
  excessBecomesUnpaid?: boolean;
  autoWeekends?: boolean;
  allowHalfDays?: boolean;
  emailNotifications?: boolean;
}

export interface TeamMemberAvailability {
  userId: number;
  userName: string;
  days: CalendarDay[];
}

export interface ReportRow {
  userId: number;
  userName: string;
  leaveYear: number;
  startDate: string;
  annualEntitlement: number;
  accrued: number;
  usedPaid: number;
  usedUnpaid: number;
  sickUsed: number;
  sickRemaining: number;
  pendingCount: number;
  approvedCount: number;
  rejectedCount: number;
}

export interface UnpaidDeduction {
  userId: number;
  userName: string;
  requestId: number;
  startDate: string;
  endDate: string;
  unpaidDays: number;
}

export interface AdminOverview {
  totalUsers: number;
  activeUsers: number;
  pendingRequests: number;
  onLeaveToday: number;
  totalUnpaidDays: number;
}

export interface Notification {
  id: number;
  userId: number;
  type: string;
  title: string;
  message: string;
  read: boolean;
  /** @nullable */
  requestId?: number | null;
  createdAt: string;
}

export interface UnreadCount {
  unread: number;
}

export type GetCalendarParams = {
start: string;
end: string;
};

export type ListMyLeaveRequestsParams = {
status?: RequestStatus;
type?: LeaveType;
};

export type GetTeamAvailabilityParams = {
start: string;
end: string;
};

export type GetTeamMemberCalendarParams = {
userId: number;
start: string;
end: string;
};

export type AdminListUsersParams = {
includeInactive?: boolean;
};

export type AdminGetUserCalendarParams = {
userId: number;
start: string;
end: string;
};

export type AdminListLeaveRequestsParams = {
status?: RequestStatus;
type?: LeaveType;
userId?: number;
};

export type AdminGetTeamAvailabilityParams = {
start: string;
end: string;
};

export type AdminGetReportParams = {
userId?: number;
leaveYear?: number;
start?: string;
end?: string;
type?: LeaveType;
status?: RequestStatus;
};

export type AdminGetUnpaidDeductionsParams = {
leaveYear?: number;
};