# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Leave & Working Days API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /calendar:
    get:
      operationId: getCalendar
      tags: [calendar]
      summary: Get the current user's calendar days within a date range
      parameters:
        - name: start
          in: query
          required: true
          schema: { type: string, format: date }
        - name: end
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Calendar days
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarDay"
  /calendar/mark:
    post:
      operationId: markCalendarDays
      tags: [calendar]
      summary: Mark one or more days as working or non-working for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarMarkInput"
      responses:
        "200":
          description: Updated calendar days
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarDay"
  /leave-requests:
    get:
      operationId: listMyLeaveRequests
      tags: [leave]
      summary: List the current user's leave requests
      parameters:
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/RequestStatus" }
        - name: type
          in: query
          required: false
          schema: { $ref: "#/components/schemas/LeaveType" }
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeaveRequest
      tags: [leave]
      summary: Create a leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveRequestInput"
      responses:
        "201":
          description: Created leave request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leave-requests/preview:
    post:
      operationId: previewLeaveRequest
      tags: [leave]
      summary: Calculate affected working days, paid and unpaid days for a prospective request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveRequestInput"
      responses:
        "200":
          description: Calculation preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeavePreview"
  /leave-requests/{id}/cancel:
    post:
      operationId: cancelLeaveRequest
      tags: [leave]
      summary: Cancel a pending leave request owned by the current user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cancelled request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /team-availability:
    get:
      operationId: getTeamAvailability
      tags: [team]
      summary: Get team availability across a date range (any authenticated member)
      parameters:
        - name: start
          in: query
          required: true
          schema: { type: string, format: date }
        - name: end
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Team availability
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMemberAvailability"
  /team/user-calendar:
    get:
      operationId: getTeamMemberCalendar
      tags: [team]
      summary: Get a specific team member's calendar days within a range (any authenticated member)
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: integer }
        - name: start
          in: query
          required: true
          schema: { type: string, format: date }
        - name: end
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Calendar days
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarDay"
...
components:
  schemas:
...
    LeaveType:
      type: string
      enum: [annual, sick]
    RequestStatus:
      type: string
      enum: [pending, approved, rejected, cancelled]
    DayStatus:
      type: string
      enum:
        - working
        - non_working
        - annual_pending
        - annual_approved
        - annual_rejected
        - sick_pending
        - sick_approved
        - sick_rejected
...
    CalendarDay:
      type: object
      required:
        - id
        - userId
        - date
        - status
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        userId: { type: integer }
        date: { type: string, format: date }
        status: { $ref: "#/components/schemas/DayStatus" }
        note: { type: ["string", "null"] }
        requestId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CalendarMarkInput:
      type: object
      required:
        - dates
        - status
      properties:
        dates:
          type: array
          items: { type: string, format: date }
        status: { $ref: "#/components/schemas/DayStatus" }
        note: { type: string }
    LeaveRequest:
      type: object
      required:
        - id
        - userId
        - type
        - startDate
        - endDate
        - workingDays
        - paidDays
        - unpaidDays
        - status
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        userId: { type: integer }
        type: { $ref: "#/components/schemas/LeaveType" }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        workingDays: { type: number }
        paidDays: { type: number }
        unpaidDays: { type: number }
        status: { $ref: "#/components/schemas/RequestStatus" }
        userNote: { type: ["string", "null"] }
        adminNote: { type: ["string", "null"] }
        approvedBy: { type: ["integer", "null"] }
        approvedAt: { type: ["string", "null"], format: date-time }
        rejectedBy: { type: ["integer", "null"] }
        rejectedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    LeaveRequestInput:
      type: object
      required:
        - type
        - startDate
        - endDate
      properties:
        type: { $ref: "#/components/schemas/LeaveType" }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        userNote: { type: string }
    LeavePreview:
      type: object
      required:
        - workingDays
        - paidDays
        - unpaidDays
        - hasUnpaid
        - availableBeforeRequest
      properties:
        workingDays: { type: number }
        paidDays: { type: number }
        unpaidDays: { type: number }
        hasUnpaid: { type: boolean }
        availableBeforeRequest: { type: number }
        sickOverAllowance: { type: boolean }
...
    TeamMemberAvailability:
      type: object
      required:
        - userId
        - userName
        - days
      properties:
        userId: { type: integer }
        userName: { type: string }
        days:
          type: array
          items:
            $ref: "#/components/schemas/CalendarDay"
...
    Notification:
      type: object
      required:
        - id
        - userId
        - type
        - title
        - message
        - read
        - createdAt
      properties:
        id: { type: integer }
        userId: { type: integer }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        read: { type: boolean }
        requestId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
...